SIMDify block-zero testing

+++ lib/sparsewrite.hh
@@ -1,8 +1,9 @@
 #include "endian.hh"
+#include <sys/types.h>
 
-bool is_zero_block(const unsigned char* data, uint_fast64_t size);
+bool is_zero_block(const unsigned char* data, size_t size);
 
 bool SparseWrite(int fd,
     const unsigned char* Buffer,
-    uint_fast64_t BufSize,
+    size_t BufSize,
     uint_fast64_t WritePos);
